# refactor: replace MersenneTwister implementation with a pure-rand based one