@matthewmayer 2024-11-30
Could there be some logic in adding pure-rand as a devDependency and adding a few tests that show that the output of our function is identical to the output from pure-rand? To guard against any possible divergence in future.