@xDivisionByZerox 2024-11-30
> Could there be some logic in adding pure-rand as a devDependency and adding a few tests that show that the output of our function is identical to the output from pure-rand? To guard against any possible divergence in future. 

That's not what we are aiming for here. Our goal is to have a faster implementation of mersenne. We only use pure-rand's implementation as a basis.